requirements.txt
Django==1.11.5
Fabric==1.14.0
Pillow==4.2.1

bpython==0.16
django-authtools==1.6.0
django-debug-toolbar==1.8
django-extensions==1.9.1
django-localflavor==1.5.2
django-redis==4.8.0
django-rq==0.9.6
djangorestframework==3.6.4
gunicorn==19.7.1
ipython[notebook]==5.3.0
pdbpp==0.9.1
pytest==3.2.2
pytest-django==3.1.2
psycopg2==2.7.3.1
pytz==2017.2
raven==6.1.0
redis==2.10.6
rq==0.8.2
Scrapy==1.5.0
